tasty_bites_api/requirements.txt
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
Django~=5.0.2
djangorestframework-simplejwt>=5.0.0
pillow>=8.4.0

python-dotenv~=1.0.1
celery~=5.2.0
redis
